Adds hypnosis spell (enthrall)
ChatGPT slopcode I spent the day on to add a hypnosis spell.

Has a 2 minute recharge and costs 7 spell points, making it among the most expensive spells to take as a magic class.

Limited to one person at a time, attempting to cast it on someone else revokes the hypnosis of the previous target.

Flavored to be taboo, unsavory magic in the hopes that people feel nasty about it.

Text includes comments about a hazy mind that doesn't allow you to recall information or locations outside of your immediate surroundings. This is a not so subtle way to make sure it isn't used by the court magus to try and make bandits spill their guts about the base location.

Tested relatively thoroughly, no visible issues.